requirements.txt
opencv-python
SpeechRecognition
numpy
librosa
openai
pydub
moviepy
streamlit
python-dotenv
